refactor(tcp): 优化TCP探测逻辑及并发处理
- 减少TCP探测次数及调整超时参数以提高效率
- 增加连续失败计数，连续两次失败且无成功探测时提前终止测试
- 设置TCP连接SetLinger为0，关闭连接时快速释放资源
- 优化批量TCP测试并发控制，使用WaitGroup和原子操作管理进度显示
- 优化带宽测量批量处理，使用同步工具提升并发安全及性能
- 调整配置文件默认值，增加最大工作线程数和带宽大小设置
- 修改输出目录写入逻辑，确保输出目录存在避免文件创建失败
- 新增docker-compose配置及完善Dockerfile以适应新的构建参数
- 移除冗余的config备份文件，改用更简洁的配置加载与初始化方式

+++ internal/model/config.go
@@ -69,15 +69,15 @@ type Availability struct {
 
 // 带宽配置
 type Bandwidth struct {
-	Enabled        bool   `json:"enabled" yaml:"enabled"`                 // 是否启用带宽控制
-	SizeMB         int    `json:"size_mb" yaml:"size_mb"`                 // 带宽大小（MB）
-	Timeout        int    `json:"timeout" yaml:"timeout"`                 // 带宽超时时间（秒）
-	Retry          int    `json:"retry" yaml:"retry"`                     // 带宽重试次数
-	RetryDelay     int    `json:"retry_delay" yaml:"retry_delay"`         // 带宽重试延迟（秒）
-	UrlTemplate    string `json:"url_template" yaml:"url_template"`       // 带宽 URL 模板
-	ProcessBuffer  int    `json:"process_buffer" yaml:"process_buffer"`   // 带宽处理缓冲区
-	ConnectTimeout int    `json:"connect_timeout" yaml:"connect_timeout"` // 带宽连接超时时间（秒）
-	MaxWorkers     int    `json:"max_workers" yaml:"max_workers"`         // 最大线程数
+	Enabled        bool    `json:"enabled" yaml:"enabled"`                 // 是否启用带宽控制
+	SizeMB         float64 `json:"size_mb" yaml:"size_mb"`                 // 带宽大小（MB）
+	Timeout        int     `json:"timeout" yaml:"timeout"`                 // 带宽超时时间（秒）
+	Retry          int     `json:"retry" yaml:"retry"`                     // 带宽重试次数
+	RetryDelay     int     `json:"retry_delay" yaml:"retry_delay"`         // 带宽重试延迟（秒）
+	UrlTemplate    string  `json:"url_template" yaml:"url_template"`       // 带宽 URL 模板
+	ProcessBuffer  int     `json:"process_buffer" yaml:"process_buffer"`   // 带宽处理缓冲区
+	ConnectTimeout int     `json:"connect_timeout" yaml:"connect_timeout"` // 带宽连接超时时间（秒）
+	MaxWorkers     int     `json:"max_workers" yaml:"max_workers"`         // 最大线程数
 }
 
 // 日志配置
